Add Direct3D11 backend and GPU profiling
Introduce Direct3D11 platform files and HLSL shaders, add Renderer3D shader/GLSL equivalents, and extend engine rendering to support Direct3D. Update BluEditorLayer to initialize and clean up GPU timestamp queries (D3D11 and GL fallbacks), measure CPU/GPU frame times, throttle perf plotting, and avoid costly per-frame ReadPixel calls. Refactor RendererAPI checks, guard OpenGL calls, switch some texture ID usages to ImGui-friendly GetImTextureID, and improve viewport input gating for the editor camera. Misc: expand .gitignore, tweak ImGui layout (imgui.ini), remove sample BlueCube scene, and small shader/renderer fixes and cleanups.

+++ Blu/engine/src/Blu/Platform/OpenGL/OpenGLFrameBuffer.h
@@ -9,7 +9,7 @@ namespace Blu
 	public:
 		OpenGLFrameBuffer(const FrameBufferSpecifications& spec);
 		virtual const FrameBufferSpecifications& GetSpecification() const override { return m_Specification; }
-		virtual uint32_t GetColorAttachmentID(uint32_t index = 0) const override { BLU_CORE_ASSERT("",index < m_ColorAttachments.size()); return m_ColorAttachments[index]; }
+		virtual uint64_t GetColorAttachmentID(uint32_t index = 0) const override { BLU_CORE_ASSERT("",index < m_ColorAttachments.size()); return (uint64_t)m_ColorAttachments[index]; }
 		virtual void Resize(uint32_t width, uint32_t height) override;
 		virtual int ReadPixel(uint32_t attachmentIndex, int x, int y) override;
 		virtual float ReadDepth(uint32_t attachmentIndex, int x, int y) override;
